#214 feat(fe): metadata 파일 분리
