+++ src/frontend/apps/web/app/stock/[stockSlug]/page.tsx
@@ -1,20 +1,12 @@
+'use client';
+
 import { ChatContainer } from '@/src/features/chat';
 import { StockDetailLayout } from '@/src/features/stock';
 import { RQProvider } from '@/src/shared';
 import { ChevronLeft } from 'lucide-react';
 import Link from 'next/link';
 
-// 주식데이터를 가지고 올수있는 방법이 필요합니다.
-
-export async function generateMetadata({ params }) {
-  const { stockSlug } = params;
-
-  return {
-    title: `${stockSlug} - 주식 정보`,
-    description: `${stockSlug}의 최신 주식 정보를 확인하세요.`,
-    keywords: `주식, ${stockSlug}, 주식 정보`,
-  };
-}
+export { generateMetadata } from './metadata';
 
 export default function StockDetailsPage({ params }) {
   const { stockSlug } = params;
